% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CastorData.R
\docType{class}
\name{CastorData-class}
\alias{CastorData-class}
\alias{CastorData}
\title{Class used to access Castor REST API.}
\format{An R6 class object.}
\usage{
CastorData
}
\description{
Class used to access Castor REST API.
}
\section{Methods}{

\itemize{
 \item \code{getStudies(): Retries all Castor studies in account.}
 \item \code{getFields(study_id, include, page): Gets all fields for a given
 study.
 }
 \item \code{getRecords(study_id): Gets all records for a given
 study.
 }
 \item \code{getStudyDataPoints(study_id, record_id,
                                filter_types): Gets all data points
 for a given study and record. Filter types may be supplied for fields types
 as a character vector of field types.
 }
 \item \code{getStudyData(study_id, filter_types): creates a data
 frame with all data points for a given study with each row representing a
 record and each column a field.
 }
 \item \code{adjustTypes = function(field_to_type, values,
                                    type_to_func): Utility method
 for casting columns to their intended type. Users can supply their own list
 of data type : function (for casting the data) mappings on instantiation of
 this class, CastorAPIWrapper.
 }
}
}

\keyword{internal}
