% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cols_to_map}
\alias{cols_to_map}
\title{cols_to_map}
\usage{
cols_to_map(dataframe, key, value)
}
\arguments{
\item{dataframe}{A data frame}

\item{key}{A string of a column name in the data frame}

\item{value}{A string of a column name in the data frame}
}
\value{
A named list, with the names from the key field and values from
the value field
}
\description{
Generates a named list based on two columns from a data frame.
}
\examples{
cols_to_map(mtcars, key = "cyl", value = "mpg")
}
