% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_value_label}
\alias{map_value_label}
\title{map_value_label}
\usage{
map_value_label(value_vector, link_list)
}
\arguments{
\item{value_vector}{The vector with values}

\item{link_list}{A list with values as names and labels as items}
}
\value{
A vector with labels
}
\description{
Maps a vector of values to labels given the value label link
}
