% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_missings}
\alias{transform_missings}
\title{transform_missings}
\usage{
transform_missings(value)
}
\arguments{
\item{value}{The value in the Castor database}
}
\value{
The value, or the respective missing value if the value was missing
}
\description{
Transforms character missing values to the respective negative numeric values
}
