% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_missings_date}
\alias{transform_missings_date}
\title{transform_missings_date}
\usage{
transform_missings_date(value)
}
\arguments{
\item{value}{The value in the Castor database}
}
\value{
The value, or the respective missing value if the value was missing
}
\description{
Transforms character missing values to the respective negative numeric values for date variables
}
