% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CastorAPIWrapper.R
\docType{class}
\name{CastorAPIWrapper-class}
\alias{CastorAPIWrapper-class}
\alias{CastorAPIWrapper}
\title{Class used to wrap Castor REST API.}
\format{An R6 class object.}
\usage{
CastorAPIWrapper
}
\description{
Class used to wrap Castor REST API.
}
\section{Methods}{

\itemize{
 \item \code{getRequest(req_path, query_data): Sends a request to the API
 given a url path and any query parameters (such as page number or include
 arguments).}
 \item \code{collectPages(request_url, page, include: }
 \item \code{getStudy(study_id): Retrieves information about given study.}
 \item \code{getStudiesPages(): Retries all Castor studies in account.}
 \item \code{getStep(study_id, step_id): Retrieves information about given study.}
 \item \code{getStepsPages(): Retries all Castor studies in account.}
 \item \code{getField(study_id, field_id, include): Gets a specified
 field.
 }
 \item \code{getFieldsPages(study_id, include, page): Gets all fields pages
 for a given study.
 }
 \item \code{getRecord(study_id, record_id): Gets a specified
 record.
 }
 \item \code{getRecordsPages(study_id: Gets all records for a given
 study.
 }
 \item \code{getStudyDataPoint(study_id, record_id,
                               field_id: Gets an individual data
 point for a given study, record and field.
 }
 \item \code{getStudyDataPointsPages(study_id, record_id,
                                     filter_types: Gets all data
 points for a given study and record. Filter types may be supplied for fields
 types as a character vector of field types.
 }
}
}

\keyword{internal}
